added discounts

+++ app-product/src/main/java/uz/pdp/appproduct/repository/DiscountRepository.java
@@ -11,8 +11,8 @@ import uz.pdp.appproduct.entity.Discount;
  */
 
 @Repository
-public interface DiscountRepository extends JpaRepository<Discount, Long> {
+public interface DiscountRepository extends JpaRepository<Discount, Integer> {
 
-    boolean existsById(Long id);
+    boolean existsById(Integer id);
 
 }
